+++ invenio_db/alembic/96e796392533_create_database_migrations.py
@@ -8,8 +8,6 @@
 
 """Create database migrations."""
 
-import sqlalchemy as sa
-from alembic import op
 
 # revision identifiers, used by Alembic.
 revision = "96e796392533"